+++ jest.setup.js
@@ -1,3 +1,0 @@
-"use strict";
-
-jest.setTimeout(30000); // in milliseconds (30 sec)