+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
-- tbd
+- Propagate subdomain to logs (if available)
 
 ## Version 1.15.12 - 2026-05-08